update TS export
