+++ packages/x-charts/src/models/seriesType/index.ts
@@ -15,11 +15,11 @@ type DefaultizedCartesianSeriesType =
 // | PieSeriesType
 type StackableSeriesType = DefaultizedBarSeriesType | DefaultizedLineSeriesType;
 
+export * from './line';
+export * from './bar';
+export * from './scatter';
 export type {
-  BarSeriesType,
-  LineSeriesType,
   // PieSeriesType,
-  ScatterSeriesType,
   AllSeriesType,
   CartesianSeriesType,
   DefaultizedCartesianSeriesType,
