+++ packages/x-charts/src/models/seriesType/line.ts
@@ -1,6 +1,15 @@
-import { CurveType } from '../../internals/getCurve';
 import { DefaultizedProps } from '../helpers';
 import { CartesianSeriesType, CommonSeriesType, DefaultizedCommonSeriesType } from './common';
+
+export type CurveType =
+  | 'catmullRom'
+  | 'linear'
+  | 'monotoneX'
+  | 'monotoneY'
+  | 'natural'
+  | 'step'
+  | 'stepBefore'
+  | 'stepAfter';
 
 export interface LineSeriesType extends CommonSeriesType, CartesianSeriesType {
   type: 'line';
